mini_projects/zsolt.komlosi/backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# LangChain & LangGraph
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# OpenAI
openai>=1.50.0

# Vector Database
qdrant-client>=1.12.0

# BM25 for hybrid search
rank-bm25>=0.2.2

# Token counting
tiktoken>=0.8.0

# HTTP Client
requests>=2.31.0
httpx>=0.28.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database (async SQLite)
aiosqlite>=0.20.0

# Jira Integration (optional)
atlassian-python-api>=3.41.0

# MCP (Model Context Protocol)
mcp>=1.0.0

# Translation
deep-translator>=1.11.4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=4.0.0
